Chore: Navbar redesign and mirigrated to mantine 7.1.1
