+++ src/components/Header/CreateUser.tsx
@@ -1,11 +1,10 @@
 import { Button, Tooltip, px } from '@mantine/core';
 import {  IconUserPlus } from '@tabler/icons-react';
 import type { FC } from 'react';
-import { useLogQueryStyles } from './styles';
+import classes from './LogQuery.module.css';
 import { useHeaderContext } from '@/layouts/MainLayout/Context';
 
 const CreateUser: FC = () => {
-	const { classes } = useLogQueryStyles();
 	const { refreshNowBtn } = classes;
 	const {
 		state: { subCreateUserModalTogle }
@@ -20,7 +19,7 @@ const CreateUser: FC = () => {
 	return (
         <Tooltip
         label={'Create new user'}
-        sx={{ color: 'white', backgroundColor: 'black' }}
+        style={{ color: 'white', backgroundColor: 'black' }}
         withArrow
         onClick={handleOpen}
         position="left">
