+++ src/components/Header/Header.module.css
@@ -1,0 +1,33 @@
+.container {
+	border-bottom: rem(1px) solid light-dark(var(--mantine-color-gray-3), var(--mantine-color-dark-4));
+}
+
+.logoContainer {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	padding: var(--mantine-spacing-md);
+	padding-right: 0;
+	min-width: var(--mantine-navbar-width);
+	height: 24px;
+}
+
+.imageSty {
+}
+
+.burgerIcon {
+}
+
+.navContainer {
+	width: calc(100%);
+	justify-content: space-between;
+}
+
+.actionBtn {
+	transition: transform 0.2s ease-in-out;
+}
+
+.actionBtn:hover {
+	transform: scale(1.3);
+	background-color: var(--mantine-colors-gray-0);
+}
