+++ src/components/Header/HeaderBreadcrumbs.tsx
@@ -1,9 +1,10 @@
 import useMountedState from '@/hooks/useMountedState';
 import { useHeaderContext } from '@/layouts/MainLayout/Context';
-import { Breadcrumbs, Text } from '@mantine/core';
+import { Breadcrumbs, Select, Text } from '@mantine/core';
 import type { FC } from 'react';
 import { useEffect } from 'react';
-import { useLogQueryStyles } from './styles';
+
+import classes from './LogQuery.module.css';
 
 type HeaderBreadcrumbs = {
 	crumbs: string[];
@@ -28,7 +29,14 @@ const HeaderBreadcrumbs: FC<HeaderBreadcrumbs> = (props) => {
 			<HomeIcon />
 			{crumbs.map((crumb) => {
 				if (crumb === 'streamName') {
-					return <Text key={crumb}>{streamName}</Text>;
+					return <Select
+					placeholder="Pick one"
+					
+					data={ ["sjs","backend", "frontend"]}
+					searchable
+					required
+				
+				/>;
 				}
 
 				return <Text key={crumb}>{crumb}</Text>;
@@ -38,7 +46,6 @@ const HeaderBreadcrumbs: FC<HeaderBreadcrumbs> = (props) => {
 };
 
 const HomeIcon: FC = () => {
-	const { classes } = useLogQueryStyles();
 	const { homeIcon } = classes;
 	return (
 		<svg className={homeIcon} xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
