+++ src/components/Header/Layout.tsx
@@ -1,22 +1,17 @@
-import { HEADER_HEIGHT } from '@/constants/theme';
-import type { HeaderProps as MantineHeaderProps } from '@mantine/core';
-import { Box, Header as MantineHeader } from '@mantine/core';
+import { Box } from '@mantine/core';
 import { FC } from 'react';
-import {  Outlet } from 'react-router-dom';
-import { useHeaderStyles } from './styles';
+import { Outlet } from 'react-router-dom';
+import classes from './Header.module.css';
 
-type HeaderProps = Omit<MantineHeaderProps, 'children' | 'height' | 'className'>;
-
-const HeaderLayout: FC<HeaderProps> = (props) => {
-	const { classes } = useHeaderStyles();
+const HeaderLayout: FC<any> = () => {
 	const { container, navContainer } = classes;
 
 	return (
-		<MantineHeader {...props} className={container} height={HEADER_HEIGHT} p={0} withBorder zIndex={100}>
+		<header className={container}>
 			<Box className={navContainer}>
 				<Outlet />
 			</Box>
-		</MantineHeader>
+		</header>
 	);
 };
