+++ src/components/Header/RefreshInterval.tsx
@@ -5,7 +5,7 @@ import { IconRefresh, IconRefreshOff } from '@tabler/icons-react';
 import ms from 'ms';
 import type { FC } from 'react';
 import { useEffect, useMemo } from 'react';
-import { useLogQueryStyles } from './styles';
+import classes from './LogQuery.module.css';
 
 const RefreshInterval: FC = () => {
 	const {
@@ -28,19 +28,19 @@ const RefreshInterval: FC = () => {
 		subRefreshInterval.set(interval);
 	};
 
-	const { classes } = useLogQueryStyles();
 	const { intervalBtn } = classes;
 
 	return (
 		<Menu withArrow>
 			<Menu.Target>
-				<Button className={intervalBtn} rightIcon={<Icon size={px('1.2rem')} stroke={1.5} />}>
+				<Button className={intervalBtn} rightSection={<Icon size={px('1.2rem')} stroke={1.5} />}>
 					<Text>{selectedInterval ? ms(selectedInterval) : 'Off'}</Text>
 				</Button>
 			</Menu.Target>
-			<Menu.Dropdown sx={{
-				zIndex: 1000,
-			}}>
+			<Menu.Dropdown
+				style={{
+					zIndex: 1000,
+				}}>
 				{REFRESH_INTERVALS.map((interval) => {
 					if (interval === selectedInterval) return null;
