+++ src/components/Header/SecondaryHeader.tsx
@@ -1,11 +1,11 @@
-import type { HeaderProps as MantineHeaderProps } from '@mantine/core';
+
 import { FC } from 'react';
 import { Route, Routes } from 'react-router-dom';
 import HeaderLayout from './Layout';
 import { ConfigHeader, LogsHeader, QueryHeader, StatsHeader, UsersManagementHeader } from './SubHeader';
 import { CONFIG_ROUTE, LOGS_ROUTE, QUERY_ROUTE, STATS_ROUTE, USERS_MANAGEMENT_ROUTE } from '@/constants/routes';
 
-type SecondaryHeaderProps = Omit<MantineHeaderProps, 'children' | 'height' | 'className'>;
+type SecondaryHeaderProps = Omit<any, 'children' | 'height' | 'className'>;
 
 const SecondaryHeader: FC<SecondaryHeaderProps> = (props) => {
 	return (
