+++ src/components/Navbar/Navbar.module.css
@@ -1,0 +1,55 @@
+.navbar {
+	height: 100vh;
+	padding: var(--mantine-spacing-md);
+	display: flex;
+	flex-direction: column;
+	border-right: rem(1px) solid light-dark(var(--mantine-color-gray-3), var(--mantine-color-dark-4));
+	transition: width 0.4s ease-in-out;
+	flex: '0 0 auto';
+}
+
+.navbarMain {
+	flex: 1;
+	margin-top: rem(50px);
+}
+
+.link {
+	height: rem(50px);
+	border-radius: var(--mantine-radius-md);
+	display: flex;
+	align-items: center;
+	color: light-dark(var(--mantine-color-gray-7), var(--mantine-color-dark-0));
+	gap: var(--mantine-spacing-sm);
+	padding: 0 var(--mantine-spacing-sm);
+
+	&:hover {
+		background-color: light-dark(var(--mantine-color-gray-0), var(--mantine-color-dark-5));
+	}
+
+	&[data-active] {
+		&,
+		&:hover {
+			background-color: var(--mantine-color-brandPrimary-light);
+			color: var(--mantine-color-brandPrimary-light-color);
+		}
+	}
+}
+.imageContainer {
+	height: rem(36px);
+	border-radius: var(--mantine-radius-md);
+	display: flex;
+	align-items: center;
+	overflow: hidden;
+	gap: var(--mantine-spacing-xs);
+	margin: 0 var(--mantine-spacing-sm);
+	padding-top: 4px;
+}
+.collapseButton {
+	position: absolute;
+	transition: all 0.4s ease-in-out;
+	top: 69px;
+	transform: translate(-50%, -50%);
+}
+.collapseButtonIcon {
+	transition: all 0.4s ease-in-out;
+}
