+++ src/layouts/FullPageLayout/FullPageLayout.module.css
@@ -1,5 +1,4 @@
 .desktop {
-	display: flex;
 	width: 100vw;
 	height: 100vh;
 	@media (max-width: 768px) {
