--- src/layouts/MainLayout/MainLayout.module.css
@@ -1,9 +1,0 @@
-.container {
-	width: 100vw;
-	min-width: 1000px;
-}
-
-.contentContainer {
-	display: flex;
-	height: calc(100vh - 64px);
-}
